Add tests and fix a couple of things
