+++ .github/workflows/code.yml
@@ -1,0 +1,52 @@
+name: Code analysis checks
+on: [push]
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [16.x]
+        python-version: [3.8]
+    env:
+        GIT_NAME: volto-export
+        NAMESPACE: "@kitconcept"
+        GIT_USER: 'kitconcept'
+    steps:
+      - name: Main checkout
+        uses: actions/checkout@v2
+
+      # node setup
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v1
+        with:
+          node-version: ${{ matrix.node-version }}
+          cache: 'yarn'
+
+      # node install
+      - name: Install dependencies
+        run: yarn
+
+      - name: Install yo and @plone/generator-volto
+        run: npm i -g yo @plone/generator-volto
+
+      - name: Extract branch name
+        run: echo "branch=${GITHUB_REF#refs/heads/}" >> $GITHUB_OUTPUT
+        id: extract_branch
+
+      - name: Setup addon testing environment
+        run: npx -p @plone/scripts addon clone .
+
+      - name: Provide ignores for eslint and prettier
+        run: cd addon-testing-project && echo -e '\nfast-blurhash.min.js' >> .eslintignore && echo -e '\nfast-blurhash.min.js' >> .prettierignore
+
+      # ESlint
+      - name: ESlint
+        run: cd addon-testing-project && yarn lint
+
+      # Stylelint
+      - name: Stylelint
+        run: cd addon-testing-project && yarn stylelint --allow-empty-input
+
+      # Prettier
+      - name: Prettier
+        run: cd addon-testing-project && yarn prettier
